Revert "recarga int"
This reverts commit cb930506ecc459db8fb02bb1c84a6a5cde5429de.

+++ include/disparo.h
@@ -6,7 +6,7 @@ class Disparo : public Ataque {
     float velocidad_ = 200.0f;  // píxeles por segundo
 
 public:
-    Disparo(int dano, float alcance, int recarga, const char* sprite, float tamanio, float duracion)
+    Disparo(int dano, float alcance, float recarga, const char* sprite, float tamanio, float duracion)
         : Ataque(dano, alcance, recarga, sprite, tamanio, duracion) {}
 
     void mover(float dt) override {
